requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
boto3==1.28.5
botocore==1.31.5
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
dataclasses-json==0.6.7
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.1.2
fastapi==0.115.8
Flask==3.0.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ads==22.1.0
google-api-core==2.24.0
google-api-python-client==2.125.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-experimental==0.3.4
langchain-text-splitters==0.3.5
langsmith==0.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.26.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.24.2
numpy==1.25.0
oauthlib==3.2.2
openai==1.61.0
openpyxl==3.1.5
orjson==3.10.15
outcome==1.2.0
packaging==24.2
pandas==2.0.3
pillow==11.1.0
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.22.3
protobuf==4.24.4
py4j==0.10.9.7
pyarrow==19.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.1.2
PySocks==1.7.1
pyspark==3.4.1
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.36.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.6.1
selenium==4.10.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.37
sqlalchemy-bigquery==1.12.1
starlette==0.45.3
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.34.0
wait-for-it==2.2.2
webdriver-manager==3.8.6
Werkzeug==3.0.0
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
